feat: Integrate spacegraph.js library and demo
Integrates the functional JavaScript library from the `old/` directory
into the main project structure under `src/`.

Key changes:
- Copied `old/spacegraph.js` to `src/spacegraph.js` and refactored it:
    - Removed comments and debug logs.
    - Improved code compactness and style.
    - Ensured robust DOM element selection and sizing.
- Created `src/spacegraph.css` from `old/index.css`.
- Developed `index.html` as a live demo and GitHub Pages site:
    - Includes an informational panel about the library.
    - Initializes SpaceGraphJS with a sample graph.
    - Uses CDN for Three.js and GSAP via import maps.
- Created `README.md` with project overview, features, and usage guide.
- Updated `package.json`:
    - Pointed "main" and "module" to `src/spacegraph.js`.
    - Updated `files` array.
    - Corrected repository URLs.
    - Aligned dependency versions in `index.html` with `package.json`.
- Removed the `old/` directory. The `depr/` directory was not removed due to sandbox issues but is not part of the active library.

The result is a functioning SpaceGraphJS library with a live demo
page, ready for deployment via GitHub Pages.
