+++ package.json
@@ -2,17 +2,17 @@
     "name": "spacegraph-zui",
     "version": "0.1.0",
     "description": "A general-purpose, extensible JavaScript library for creating Zooming User Interfaces (ZUI) with 2D and 3D visualizations.",
-    "main": "dist/spacegraph.umd.js",
-    "module": "dist/spacegraph.esm.js",
-    "types": "dist/spacegraph.d.ts",
+    "main": "src/spacegraph.js",
+    "module": "src/spacegraph.js",
+    "types": "",
     "files": [
-        "dist",
+        "src",
         "README.md",
-        "CORE_CONCEPTS.md",
-        "index.css"
+        "index.html",
+        ".github"
     ],
     "scripts": {
-        "build": "vite build && tsc",
+        "build": "echo \"Build script not configured for src/ direct usage yet. Using Vite for dev only.\" && vite build && tsc",
         "test": "vitest",
         "docs:build": "jsdoc -c jsdoc.conf.json",
         "docs:view": "serve ./docs/api -l 8080",
@@ -29,7 +29,7 @@
     "type": "module",
     "repository": {
         "type": "git",
-        "url": "git+https://github.com/automenta/spacegraphjs.git"
+        "url": "git+https://github.com/TTime/spacegraphjs.git"
     },
     "keywords": [
         "zui",
@@ -46,9 +46,9 @@
         "library"
     ],
     "bugs": {
-        "url": "https://github.com/automenta/spacegraphjs/issues"
+        "url": "https://github.com/TTime/spacegraphjs/issues"
     },
-    "homepage": "https://github.com/automenta/spacegraphjs#readme",
+    "homepage": "https://github.com/TTime/spacegraphjs#readme",
     "dependencies": {
         "gsap": "^3.13.0",
         "three": "^0.177.0"
